+++ src/dis6/FastEntityStatePdu.cpp
@@ -601,7 +601,7 @@ void FastEntityStatePdu::unmarshal(DataStream& dataStream)
     dataStream >> _capabilities;
 
      _articulationParameters.clear();
-     for(size_t idx = 0; idx < _numberOfArticulationParameters; idx++)
+     for(int8_t idx = 0; idx < _numberOfArticulationParameters; idx++)
      {
         ArticulationParameter x;
         x.unmarshal(dataStream);
